requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.28
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-jose[cryptography]>=3.3.0
pydantic>=2.6.3
pydantic-settings>=2.2.1
email-validator>=2.1.1
python-multipart>=0.0.9
joblib>=1.3.2
scikit-learn>=1.4.1.post1
numpy>=1.26.4
reportlab>=4.0.0
pyserial>=3.5
slowapi>=0.1.9
redis>=5.0.1
httpx>=0.27.0
qrcode
pillow
